fs: move linux watcher to internal/fs/watch
