+++ lib/fs.js
@@ -57,7 +57,7 @@ const {
 
 const pathModule = require('path');
 const { isArrayBufferView } = require('internal/util/types');
-const linuxWatcher = require('internal/fs/linux_watcher');
+const linuxWatcher = require('internal/fs/watch/linux');
 
 // We need to get the statValues from the binding at the callsite since
 // it's re-initialized after deserialization.
