+++ lib/internal/fs/watch/linux.js
@@ -3,10 +3,9 @@
 const { EventEmitter } = require('events');
 const path = require('path');
 const { SafeMap, Symbol, StringPrototypeStartsWith } = primordials;
-const { validateObject } = require('internal/validators');
+const { validateObject, validateString } = require('internal/validators');
 const { kEmptyObject } = require('internal/util');
 const { ERR_FEATURE_UNAVAILABLE_ON_PLATFORM } = require('internal/errors');
-const { Dirent, UV_DIRENT_DIR } = require('internal/fs/utils');
 
 const kFSWatchStart = Symbol('kFSWatchStart');
 
@@ -23,20 +22,25 @@ function lazyLoadFsSync() {
   return internalSync;
 }
 
-function traverse(dir, files = new SafeMap()) {
-  const { readdirSync } = lazyLoadFsSync();
+async function traverse(dir, parent = undefined, files = new SafeMap()) {
+  const { readdir, stat } = lazyLoadFsPromises();
 
-  const filenames = readdirSync(dir, { withFileTypes: true });
+  const filenames = await readdir(dir, { withFileTypes: true });
 
-  files.set(dir, new Dirent(dir, UV_DIRENT_DIR));
+  if (parent !== undefined) {
+    files.set(dir, parent);
+  } else {
+    // Optimization: Only call `stat` on the root element.
+    files.set(dir, await stat(dir));
+  }
 
-  for (const file of filenames) {
+  for await (const file of filenames) {
     const f = path.join(dir, file.name);
 
     files.set(f, file);
 
     if (file.isDirectory()) {
-      traverse(f, files);
+      await traverse(f, file, files);
     }
   }
 
@@ -49,14 +53,6 @@ class FSWatcher extends EventEmitter {
   #files = new SafeMap();
   #rootPath = path.resolve();
 
-  /**
-   * @param {{
-   *   persistent?: boolean;
-   *   recursive?: boolean;
-   *   encoding?: string;
-   *   signal?: AbortSignal;
-   *   }} [options]
-   */
   constructor(options = kEmptyObject) {
     super();
 
@@ -75,7 +71,13 @@ class FSWatcher extends EventEmitter {
     this.emit('close');
   }
 
+  /**
+   * @param {string} file
+   * @return {string}
+   */
   #getPath(file) {
+    validateString(file, 'file');
+
     if (file === this.#rootPath) {
       return this.#rootPath;
     }
@@ -84,6 +86,8 @@ class FSWatcher extends EventEmitter {
   }
 
   #unwatchFolder(file) {
+    validateString(file, 'file');
+
     const { unwatchFile } = lazyLoadFsSync();
 
     for (const filename of this.#files.keys()) {
@@ -94,7 +98,9 @@ class FSWatcher extends EventEmitter {
   }
 
   async #watchFolder(folder) {
-    const { opendir, stat } = lazyLoadFsPromises();
+    validateString(folder, 'folder');
+
+    const { opendir } = lazyLoadFsPromises();
 
     try {
       const files = await opendir(folder);
@@ -107,12 +113,10 @@ class FSWatcher extends EventEmitter {
         }
 
         if (!this.#files.has(f)) {
-          const fileStats = await stat(f);
-
-          this.#files.set(f, fileStats);
+          this.#files.set(f, file);
           this.emit('change', 'rename', this.#getPath(f));
 
-          if (fileStats.isDirectory()) {
+          if (file.isDirectory()) {
             await this.#watchFolder(f);
           } else {
             this.#watchFile(f);
@@ -124,10 +128,9 @@ class FSWatcher extends EventEmitter {
     }
   }
 
-  /**
-   * @param {string} file
-   */
   #watchFile(file) {
+    validateString(file, 'file');
+
     const { watchFile } = lazyLoadFsSync();
 
     if (this.#closed) {
@@ -155,21 +158,18 @@ class FSWatcher extends EventEmitter {
           this.#unwatchFolder(file);
         }
       } else if (statWatcher.isDirectory()) {
-        this.#watchFolder(file);
         this.emit('change', 'change', this.#getPath(file));
+        this.#watchFolder(file);
       } else {
         this.emit('change', 'change', this.#getPath(file));
       }
     });
   }
 
-  /**
-   * @param {string | Buffer | URL} filename
-   */
-  [kFSWatchStart](filename) {
+  async [kFSWatchStart](filename) {
     this.#rootPath = filename;
     this.#closed = false;
-    this.#files = traverse(filename);
+    this.#files = await traverse(filename);
 
     for (const f of this.#files.keys()) {
       this.#watchFile(f);
